make LabelDropdown a generic component
